Update README.md to be less shell dependant
Removed the need for pipe (|), command expansion with backticks (`), and
the which Bash function.

Typically you want to stay away from using `which` in a script and
instead use something like `type -p` since it is more universal.  Since
`brew` consistently installs commands to `/usr/local/bin` we can provide
an absolute path here instead.

The `grep` command can take in a file which allows us to remove the need
to run `cat` and remove the need for a pipe.

This has the added benefit of working even if someone uses a shell other
than Bash, such as Fish.